New HUD Prefab created [Read Description]
New prefab for hud created. PLEASE LISTEN WHEN GETTING CODE TO WORK WITH IT:
Under the empty game object "AttackButtons", the buttons there switch out with each other based on the conditions of the character. So if you have ammo, it's attack. If you're empty, it's reload. If you selected attack but haven't picked a target and want to cancel, do cancel attack.

The thing in the center of the middle hud is to tell the player what state they're in. By default, it should be the Moving one. If the player clucked Attack, it should switch to the Attack one. And when the player cancels attack or shoots a target, it should go back to Moving until the player hits the attack button/hotkey again.

Ammo has 2 text objects to edit. One is the current ammo in gun and the other is the maximum in the magazine. The third text there is just a / to differentiate and doesn't need to be touched.

Health bar should work the same way it does in the previous builds. It just looks different. 

Grenade button should be SetActive(false) until the player picks up the grenade. Since there's no code in the game for the grenade yet, no need for it to be visible.

Good luck with the coding, whoever does it.