Better Looking Win Screen
Created a better looking win screen but it needs the coders to come in an touch up on the original code of how it worked. 

Created empty object in the canvas that has 2 text objects, one for blue win, one for red win, and simply needs to be coded so one turns on and other stays inactive. Did the same for a model of the winning dog dancing, one for red, one for blue win. I don't know how the coding for telling the winner works so i couldn't get it to actually do these thing.

Coding aside, I also need the animation for the dancing to be added in. Searched through the asset folder and the google drive but couldn't find the animation, just a gif of it.