Team Name Display more prominent, Custom Cursor, Some Script Updating
Created new UI to display what team is active currently. Edited the script that does the display to fit this new display look.

Created a custom cursor in the form of a bone.

Setup the new Team Display to show all team members and how many turns they have left but didn't do any scripting to make that part of the UI work, someone else will have to take that job. It seems to be listed as not urgent in the programming tab of Hack'n'plan.